refactor: 워크스페이스 설정 패널 컴포넌트 분리
